feat(json): dynamic backend; `sonic` -> `go-json` (fallback)
