+++ pkg/utils/json/config.go
@@ -1,0 +1,6 @@
+package json
+
+import "github.com/bytedance/sonic"
+
+// Config is a configuration for the JSON package.
+type Config = sonic.Config
